Create custom hook to reuse existing OfferContentCTA  when an EventCard is clicked
