+++ src/features/offer/components/OfferCTAButton/OfferCTAButton.tsx
@@ -1,14 +1,12 @@
-import { useFocusEffect, useRoute } from '@react-navigation/native'
+import { useFocusEffect } from '@react-navigation/native'
 import React, { FunctionComponent, useCallback } from 'react'
 
 import { OfferResponse } from 'api/gen'
 import { useAuthContext } from 'features/auth/context/AuthContext'
-import { StepperOrigin, UseRouteType } from 'features/navigation/RootNavigator/types'
+import { useOfferCTAButton } from 'features/offer/components/OfferCTAButton/useOfferCTAButton'
 import { StickyBookingButton } from 'features/offer/components/StickyBookingButton/StickyBookingButton'
 import { getIsFreeDigitalOffer } from 'features/offer/helpers/getIsFreeDigitalOffer/getIsFreeDigitalOffer'
-import { useCtaWordingAndAction } from 'features/offer/helpers/useCtaWordingAndAction/useCtaWordingAndAction'
 import { Subcategory } from 'libs/subcategories/types'
-import { useBookOfferModal } from 'shared/offer/helpers/useBookOfferModal'
 
 type Props = {
   offer: OfferResponse
@@ -21,29 +19,11 @@ export const OfferCTAButton: FunctionComponent<Props> = ({
   subcategory,
   trackEventHasSeenOfferOnce,
 }) => {
-  const route = useRoute<UseRouteType<'Offer'>>()
-  const from = route.params?.from
-  const searchId = route.params?.searchId
-  const openModalOnNavigation = route.params?.openModalOnNavigation
+  const { ctaWordingAndAction, showOfferModal, CTAOfferModal, openModalOnNavigation } =
+    useOfferCTAButton(offer, subcategory)
+
   const { isLoggedIn } = useAuthContext()
   const isFreeDigitalOffer = getIsFreeDigitalOffer(offer)
-  const {
-    wording,
-    onPress: onPressCTA,
-    navigateTo,
-    externalNav,
-    modalToDisplay,
-    isEndedUsedBooking,
-    bottomBannerText,
-    isDisabled,
-  } = useCtaWordingAndAction({ offer, from, searchId, subcategory }) ?? {}
-
-  const { OfferModal: CTAOfferModal, showModal: showOfferModal } = useBookOfferModal({
-    modalToDisplay,
-    offerId: offer.id,
-    isEndedUsedBooking,
-    from: StepperOrigin.OFFER,
-  })
 
   useFocusEffect(
     useCallback(() => {
@@ -54,22 +34,10 @@ export const OfferCTAButton: FunctionComponent<Props> = ({
     }, [trackEventHasSeenOfferOnce, openModalOnNavigation, showOfferModal])
   )
 
-  const onPress = () => {
-    onPressCTA?.()
-    showOfferModal()
-  }
-
   return (
     <React.Fragment>
       <StickyBookingButton
-        ctaWordingAndAction={{
-          wording,
-          onPress,
-          navigateTo,
-          externalNav,
-          isDisabled,
-          bottomBannerText,
-        }}
+        ctaWordingAndAction={ctaWordingAndAction}
         isFreeDigitalOffer={isFreeDigitalOffer}
         isLoggedIn={isLoggedIn}
       />
